requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-ckeditor==6.7.1  # This version uses CKEditor 4.25.0-lts
django-cors-headers==4.3.1
python-dotenv==1.0.0
redis==5.0.1
django-redis==5.4.0
googletrans==3.1.0a0
pytest==7.4.4
pytest-django==4.7.0
flake8==7.0.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
